my_blog/requirements.txt
asgiref==3.8.1
autopep8==1.6.0
Django==5.1.6
django-crispy-forms==1.14.0
pillow==11.1.0
pycodestyle==2.8.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
